Restructure Pinia store layer (Theme B, P0+P1)
Move all per-report data stores from src/api/stores/ into src/stores/report/,
making src/api/ a pure data-layer (models, utils, worker — no Pinia). App-level
stores (reports, breadcrumb) stay at src/stores/ top-level.

Replace the misnamed api.store with a useReportLoader composable: it orchestrates
the sequential metadata->file->kgram->pair hydration and returns local reactive
loading/loadingText/error state to analysis.vue. Each data store now reads its CSV
base URL directly from useAppMode().dataUrl.

Extract cutoff and isAnonymous into a new settings.store so components that only
need the cutoff slider do not have to import the hydration coordinator.
