--- web/src/api/stores/index.ts
@@ -1,5 +1,0 @@
-export * from "./file.store";
-export * from "./kgram.store";
-export * from "./metadata.store";
-export * from "./pair.store";
-export * from "./api.store";
