+++ web/src/components/submission/SubmissionHistogram.vue
@@ -10,7 +10,7 @@
 
 <script lang="ts" setup>
 import { File } from "@/api/models";
-import { useFileStore } from "@/api/stores";
+import { useFileStore } from "@/stores/report";
 import { getLargestFieldOfScore } from "@/util/FileInterestingness";
 import { storeToRefs } from "pinia";
 import { computed } from "vue";
