+++ web/src/composables/index.ts
@@ -1,5 +1,6 @@
 export * from "./usePartialLegend";
 export * from "./useCluster";
 export * from "./useAppMode";
+export * from "./useReportLoader";
 
 export * from "./d3/useD3Tooltip";
