+++ web/src/stores/report/file.store.ts
@@ -1,9 +1,10 @@
 import { defineStore } from "pinia";
 import { shallowRef, computed, nextTick, watch, ref, ComputedRef } from "vue";
 import { File, Label, Legend, Pair } from "@/api/models";
-import { useApiStore, usePairStore } from "@/api/stores";
+import { useSettingsStore, usePairStore } from "@/stores/report";
+import { useAppMode } from "@/composables";
 import { names, animals, uniqueNamesGenerator } from "unique-names-generator";
-import { commonFilenamePrefix, parseCsv } from "../utils";
+import { commonFilenamePrefix, parseCsv } from "@/api/utils";
 import {
   FileInterestingnessCalculator,
   FileScoring,
@@ -15,8 +16,9 @@ import {
  */
 export const useFileStore = defineStore("file", () => {
   // Stores
-  const apiStore = useApiStore();
+  const settingsStore = useSettingsStore();
   const pairStore = usePairStore();
+  const { dataUrl } = useAppMode();
 
   // State
   const hydrated = shallowRef(false);
@@ -117,7 +119,7 @@ export const useFileStore = defineStore("file", () => {
   }
 
   async function fetch(
-    url: string = apiStore.url + "/files.csv"
+    url: string = dataUrl.value + "/files.csv"
   ): Promise<any[]> {
     return await parseCsv(url);
   }
@@ -242,9 +244,7 @@ export const useFileStore = defineStore("file", () => {
 
   // Anonymize the data.
   function anonymize(): void {
-    apiStore.loading = true;
-    apiStore.loadingText = "Anonymizing files...";
-    const isAnonymous = apiStore.isAnonymous;
+    const isAnonymous = settingsStore.isAnonymous;
 
     // Update the files.
     for (const file of Object.values(filesById.value)) {
@@ -275,7 +275,6 @@ export const useFileStore = defineStore("file", () => {
       filesById.value = { ...filesById.value };
       pairStore.pairs = { ...pairStore.pairs };
       legend.value = { ...legend.value };
-      apiStore.loading = false;
     });
   }
