+++ web/src/views/analysis/clusters.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script lang="ts" setup>
-import { usePairStore } from "@/api/stores";
+import { usePairStore } from "@/stores/report";
 import { storeToRefs } from "pinia";
 
 const pairStore = usePairStore();
