+++ web/src/views/analysis/pairs.vue
@@ -30,7 +30,7 @@
 </template>
 
 <script lang="ts" setup>
-import { usePairStore } from "@/api/stores";
+import { usePairStore } from "@/stores/report";
 import { useRouteQuery } from "@vueuse/router";
 
 const pairStore = usePairStore();
